+++ src/app/order/[orderId]/page.tsx
@@ -8,13 +8,12 @@ import { Badge } from '@/app/_components/ui/Badge';
 import { Button } from '@/app/_components/ui/Button';
 import OrderCard from '@/app/order/[orderId]/_components/OrderCard';
 import Layout from '@/components/layout';
+import { IRecruitItem } from '@/domain/order';
 import {
   useOrderRecruitGroupExit,
   useOrderRecruitGroupJoin,
 } from '@/libs/hooks/services/mutations/order';
 import { useRecruitQuery } from '@/libs/hooks/services/queries/order';
-
-import { IRecruitItem } from '../../../types/order';
 
 type Props = {
   params: {
